LayoutTests/imported/w3c/web-platform-tests/css/css-backgrounds/animations/border-image-slice-composition-expected.txt

CONSOLE MESSAGE: line 13: ReferenceError: Can't find variable: test_composition

Harness Error (FAIL), message = ReferenceError: Can't find variable: test_composition
